Timer 3.0
solve the 2.0 problem: when CPU don't work, neither the alarm works.
using alarmmanager and Attention: when Using with MIUI, can't timing
good in 5 mins. Because of the System Optimise the code to save
power.When time is less than 5 min, it don't work. This release also
using the Foreground  service